Factored SQL comment-getting functions out into a module.
